refactor: move parseSemver from src/utils.ts to src/semver.ts
