+++ src/core.ts
@@ -1,9 +1,8 @@
 import { ModuleJson } from "https://deno.land/x/deno_graph@0.55.0/mod.ts";
-import { createUrl, type Maybe, parseSemVer } from "./utils.ts";
+import { createUrl, type Brand, type Maybe } from "./utils.ts";
+import { parseSemVer } from "./semver.ts";
 
-// Ref: https://semver.org/#is-there-a-suggested-regular-expression-regex-to-check-a-semver-string
-const SEMVER_REGEXP =
-  /@v?(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(?:-((?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?/g;
+export type Specifier = Brand<string, "specifier">;
 
 export interface DependencyProps {
   specifier: string;
@@ -21,14 +20,15 @@ export function parseDependencyProps(
     return;
   }
   const body = url.hostname + url.pathname;
-  const match = body.match(SEMVER_REGEXP);
-  if (!match) {
+  const semver = parseSemVer(specifier);
+  if (!semver) {
     // The specifier does not contain a semver.
     return;
   }
-  const name = body.split(match[0])[0];
-  const path = body.slice(name.length + match[0].length);
-  return { specifier, name, version: match[0].slice(1), path };
+  const atSemver = "@" + semver;
+  const name = body.split(atSemver)[0];
+  const path = body.slice(name.length + atSemver.length);
+  return { specifier, name, version: semver, path };
 }
 
 type DependencyJson = NonNullable<ModuleJson["dependencies"]>[number];
