+++ src/semver_test.ts
@@ -1,6 +1,6 @@
 import { describe, it } from "https://deno.land/std@0.202.0/testing/bdd.ts";
 import { assertEquals } from "https://deno.land/std@0.202.0/assert/mod.ts";
-import { parseSemVer } from "./utils.ts";
+import { parseSemVer } from "./semver.ts";
 
 describe("parseSemVer", () => {
   it("https://deno.land/std", () =>
